+++ examples/spot/public_market_data/V3Depth/V3Depth.go
@@ -6,14 +6,14 @@ import (
 )
 
 /*
-GET /account/v1/currencies
-Doc: https://developer-pro.bitmart.com/en/spot/#get-currencies
+GET /spot/quotation/v3/books
+Doc: https://developer-pro.bitmart.com/en/spot/#get-depth-v3
 */
 func main() {
 	client := bitmart.NewClient(bitmart.Config{TimeoutSecond: 5})
 
-	// Get Currencies (V1)
-	var ac, err = client.GetSpotCurrencies()
+	// Get Depth (V3)
+	var ac, err = client.GetSpotV3Book("BTC_USDT", 10)
 	if err != nil {
 		log.Panic(err)
 	} else {