+++ examples/spot/public_market_data/V3Ticker/V3Ticker.go
@@ -6,14 +6,14 @@ import (
 )
 
 /*
-GET /account/v1/currencies
-Doc: https://developer-pro.bitmart.com/en/spot/#get-currencies
+GET /spot/quotation/v3/ticker
+Doc: https://developer-pro.bitmart.com/en/spot/#get-ticker-of-a-trading-pair-v3
 */
 func main() {
 	client := bitmart.NewClient(bitmart.Config{TimeoutSecond: 5})
 
-	// Get Currencies (V1)
-	var ac, err = client.GetSpotCurrencies()
+	// Get Ticker of a Trading Pair (V3)
+	var ac, err = client.GetSpotV3Ticker("BTC_USDT")
 	if err != nil {
 		log.Panic(err)
 	} else {