+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 Changelog
 =========================
 
+
+
+###### 2023-09-08
+- New endpoints for API Futures Trading
+  - <code>/contract/private/get-open-orders</code>Get All Open Orders (KEYED)
+  - <code>/contract/private/submit-leverage</code>Submit Leverage (SIGNED)
+- New endpoints for API Spot Market
+  - <code>/spot/quotation/v3/tickers</code> Get Ticker of All Pairs (V3)
+    <code>/spot/quotation/v3/ticker</code> Get Ticker of a Trading Pair(V3)
+    <code>/spot/quotation/v3/lite-klines</code> Get Latest K-Line (V3)
+    <code>/spot/quotation/v3/klines</code> Get History K-Line (V3)
+    <code>/spot/quotation/v3/books</code> Get Depth(V3)
+    <code>/spot/quotation/v3/trades</code> Get Recent Trades(V3)
+
+---
+
 ###### 2023-06-09
 - New endpoints for API Futures Trading
     - <code>/contract/private/submit-plan-order</code>Applicable for placing contract plan orders
@@ -19,6 +35,8 @@ Changelog
   - <code>/spot/v3/orders</code>
   - <code>/spot/v2/trades</code>
   - <code>/contract/v1/tickers</code>
+
+---
 
 ###### 2022-11-8
 - New endpoints for Contract Market